requirements.txt
appdirs==1.4.3
configparser==3.5.0
Django==2.0.1
django-ajax-selects==1.6.1
django-appconf==1.0.2
django-formtools==2.1
django-nocaptcha-recaptcha==0.0.20
flake8==3.5.0
mccabe==0.6.1
num2words==0.5.6
packaging==16.8
pdfrw==0.1
pycodestyle==2.3.1
pyflakes==1.6.0
pyparsing==2.2.0
pytz==2017.3
reportlab==3.0
six==1.11.0
